manage the state of the canvas
