+++ client/components/DrawingTool.js
@@ -5,31 +5,25 @@ export default class DrawingTool extends React.Component {
   constructor(props) {
     super(props)
 
-    // this.drawOnCanvas = this.drawOnCanvas.bind(this)
+    this.drawOnCanvas = this.drawOnCanvas.bind(this)
   }
 
-  componentDidMount() {
-    // this.props.canvas.isDrawingMode = false
-    // this.props.canvas.freeDrawingBrush.width = 15
-    // this.props.canvas.freeDrawingBrush.color = '#00aeff'
+  drawOnCanvas() {
+    console.log('drawing....')
+    this.props.canvas.isDrawingMode = !this.props.canvas.isDrawingMode
+    if (this.props.canvas.isDrawingMode) {
+      document.getElementById('draw-button').innerHTML = 'Draw Mode ON'
+    } else {
+      document.getElementById('draw-button').innerHTML = 'Draw Mode OFF'
+    }
   }
-
-  // drawOnCanvas() {
-  //   console.log('drawing....')
-  //   this.props.canvas.isDrawingMode = !this.props.canvas.isDrawingMode
-  //   if (this.props.canvas.isDrawingMode) {
-  //     document.getElementById('draw-button').innerHTML = 'Draw Mode ON'
-  //   } else {
-  //     document.getElementById('draw-button').innerHTML = 'Draw Mode OFF'
-  //   }
-  // }
 
   render() {
     return (
       <Button
         id="draw-button"
         onClick={() => {
-          this.props.drawOnCanvas()
+          this.drawOnCanvas()
         }}
       >
         Draw Mode : OFF
