+++ client/components/HomeView.js
@@ -11,7 +11,6 @@ import Grid from '@material-ui/core/Grid'
 
 import Button from '@material-ui/core/Button'
 import DrawingTool from './DrawingTool'
-import {red} from '@material-ui/core/colors'
 
 const styles = theme => ({
   root: {
@@ -35,23 +34,11 @@ class HomeView extends React.Component {
     super(props)
     this.addToCanvas = this.addToCanvas.bind(this)
     this.clearEl = this.clearEl.bind(this)
-    // this.state = {
-    //   canvas: new fabric.Canvas(document.getElementById('my-canvas'))
-    // }
-    // this.state = {
-    //   canvas: null
-    // }
   }
 
   componentDidMount() {
     this.props.fetchStickers()
-    // this.setState({
-    //   canvas: new fabric.Canvas('my-canvas')
-    // })
     this.canvas = new fabric.Canvas('my-canvas')
-    // this.JSON = JSON.stringify(this.canvas)
-    // console.log('my canvas', this.state.canvas)
-    // this.JSON = JSON.stringify(this.canvas)
   }
 
   addToCanvas(sticker) {
@@ -67,35 +54,6 @@ class HomeView extends React.Component {
       },
       {crossOrigin: 'Anonymous'}
     )
-  }
-
-  exportFile() {
-    // 1) grab canvas 'DOM' element
-    // 2) call ToBlob function on the canvas DOM and SaveAs.'file_name.jpeg'
-    const exportCanvas = document.getElementById('my-canvas')
-    exportCanvas.toBlob(function(blob) {
-      saveAs(blob, 'eureka_img.jpeg')
-    })
-  }
-
-  saveFile() {
-    // 1) grab the canvas and convert to JSON
-    // 2) make an axios request to add this JSON to DB
-    console.log('saving file to DB')
-    let canvasJSON = JSON.stringify(this.canvas.toDatalessJSON())
-    console.log('JSON', canvasJSON)
-    // console.log('toObject', this.canvas.toObject());//logs canvas as an object
-    // console.log('toSVG', this.canvas.toSVG());//logs the SVG representation of canvas
-  }
-
-  drawOnCanvas() {
-    console.log('drawing....')
-    this.canvas.isDrawingMode = !this.canvas.isDrawingMode
-    if (this.canvas.isDrawingMode) {
-      document.getElementById('draw-button').innerHTML = 'Draw Mode : ON'
-    } else {
-      document.getElementById('draw-button').innerHTML = 'Draw Mode : OFF'
-    }
   }
 
   clearEl() {
@@ -117,10 +75,7 @@ class HomeView extends React.Component {
               />
             </Paper>
 
-            <DrawingTool
-              canvas={this.canvas}
-              drawOnCanvas={this.drawOnCanvas}
-            />
+            <DrawingTool canvas={this.canvas} />
 
             <Button
               onClick={() => {
